ARM: sunxi: Adapt to the unsigned long change for get_ram_size()
get_ram_size() now properly returns unsigned long. This patch adapts
sunxi to also work with unsigned long's.


Conflicts:
	board/sunxi/board.c

+++ arch/arm/include/asm/arch-sunxi/dram.h
@@ -185,7 +185,7 @@ struct dram_para {
 
 #define DRAM_CSEL_MAGIC 0x16237495
 
-int sunxi_dram_init(void);
-int dramc_init(struct dram_para *para);
+unsigned long sunxi_dram_init(void);
+unsigned long dramc_init(struct dram_para *para);
 
 #endif /* _SUNXI_DRAM_H */
